Creating Unity Player Analytics
This is a project where you can track any movement in the project. So the way that it works is for the analytics there are three types of scripts, two of them work together and one of the scripts works independently.

So for the two scripts that work together one of the scripts is called "playerSavePosData" and it does exactly that it saves the player's position data. The Second script is called "PlayerLoadPosData" and again it does exactly that it load's the player's position data from the saved file which is located in the Assets folder and it is called "LogPlayerPosData"

The third script does something a bit different however it is similar to the previous script. So it records the player's position and then uses that data to draw lines and boxes in the editor, in the path that the player travelled in. One thing that this script does not do is save's the movement data, so when you stop playing the game the data from the editor will disappear.

In this commit, I have added all the functionality of the player movement tracking system, which I have been using for analytics.

In future Commits/iterations of this project, I will be refining the project and will possibly be making some improvements to the code and the project environment.